Add FileService for reading and saving data from/to a file
Added  FileService, a service responsible for reading data from and saving data to a file. The service uses Newtonsoft.Json for JSON serialization and utilizes IConfiguration to access configuration values, such as the path to the file to read from and save to. The service supports reading the content of the file as well as saving a list of ToDoItems to the specified file in JSON format.